+++ src/main/java/com/map/gaja/client/application/ClientUpdater.java
@@ -1,4 +1,4 @@
-package com.map.gaja.client.apllication;
+package com.map.gaja.client.application;
 
 import com.map.gaja.client.domain.model.Client;
 import com.map.gaja.client.domain.model.ClientAddress;
@@ -7,7 +7,7 @@ import com.map.gaja.client.domain.model.ClientLocation;
 import com.map.gaja.client.presentation.dto.request.NewClientRequest;
 import com.map.gaja.client.presentation.dto.subdto.StoredFileDto;
 
-import static com.map.gaja.client.apllication.ClientConvertor.dtoToVo;
+import static com.map.gaja.client.application.ClientConvertor.dtoToVo;
 
 public class ClientUpdater {
     /**
