+++ src/main/java/com/map/gaja/client/application/validator/ClientRequestValidator.java
@@ -1,4 +1,4 @@
-package com.map.gaja.client.apllication.validator;
+package com.map.gaja.client.application.validator;
 
 import com.map.gaja.client.infrastructure.file.FileValidator;
 import com.map.gaja.client.infrastructure.file.exception.FileNotAllowedException;
