+++ src/main/java/com/map/gaja/client/infrastructure/location/exception/NotExcelUploadException.java
@@ -1,4 +1,4 @@
-package com.map.gaja.global.location.exception;
+package com.map.gaja.client.infrastructure.location.exception;
 
 import com.map.gaja.global.exception.BusinessException;
 import org.springframework.http.HttpStatus;
