+++ src/main/java/com/map/gaja/global/config/SecurityConfig.java
@@ -1,28 +1,25 @@
 package com.map.gaja.global.config;
 
 import com.map.gaja.global.filter.RemoveAnonymousAuthenticationFilter;
-import com.map.gaja.global.oauth2.application.Oauth2UserService;
-import com.map.gaja.global.oauth2.handler.Oauth2FailureHandler;
-import com.map.gaja.global.oauth2.handler.Oauth2SuccessHandler;
+import com.map.gaja.user.infrastructure.oauth2.OAuth2WebService;
+import com.map.gaja.user.infrastructure.oauth2.handler.OAuth2WebFailureHandler;
+import com.map.gaja.user.infrastructure.oauth2.handler.OAuth2WebSuccessHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.AnonymousAuthenticationFilter;
 import org.springframework.security.web.authentication.HttpStatusEntryPoint;
-import org.springframework.security.web.firewall.DefaultHttpFirewall;
-import org.springframework.security.web.firewall.HttpFirewall;
 
 @EnableWebSecurity
 @RequiredArgsConstructor
 public class SecurityConfig {
-    private final Oauth2UserService oauth2UserService;
-    private final Oauth2SuccessHandler oauth2SuccessHandler;
-    private final Oauth2FailureHandler oauth2FailureHandler;
+    private final OAuth2WebService OAuth2WebService;
+    private final OAuth2WebSuccessHandler OAuth2WebSuccessHandler;
+    private final OAuth2WebFailureHandler OAuth2WebFailureHandler;
 
     @Value("${management.endpoints.web.base-path}")
     private String monitoringEndPoint;
@@ -44,10 +41,10 @@ public class SecurityConfig {
                 .addFilterBefore(new RemoveAnonymousAuthenticationFilter(), AnonymousAuthenticationFilter.class)
                 .formLogin().disable()
                 .oauth2Login()
-                .userInfoEndpoint().userService(oauth2UserService) //oauth2 로그인 후 처리 로직
+                .userInfoEndpoint().userService(OAuth2WebService) //oauth2 로그인 후 처리 로직
                 .and()
-                .successHandler(oauth2SuccessHandler) //웹에서 oauth2 로그인 성공할 경우
-                .failureHandler(oauth2FailureHandler) //웹에서 oauth2 로그인 실패할 경우(회원탈퇴 유저가 로그인하면 예외)
+                .successHandler(OAuth2WebSuccessHandler) //웹에서 oauth2 로그인 성공할 경우
+                .failureHandler(OAuth2WebFailureHandler) //웹에서 oauth2 로그인 실패할 경우(회원탈퇴 유저가 로그인하면 예외)
                 .and()
                 .exceptionHandling()
                 .authenticationEntryPoint(new HttpStatusEntryPoint(HttpStatus.UNAUTHORIZED)); //인증되지 없을 경우 로그인 창으로 이동하는데 앱에서는 401으로 응답해주기 위해서 401으로 통일함
