+++ src/test/java/com/map/gaja/user/infrastructure/Oauth2ClientTest.java
@@ -1,6 +1,6 @@
 package com.map.gaja.user.infrastructure;
 
-import org.junit.jupiter.api.Test;
+import com.map.gaja.user.infrastructure.oauth2.OAuth2Appservice;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -8,10 +8,10 @@ class Oauth2ClientTest {
 
 //    @Test
     void oauth2Login() {
-        Oauth2Client oauth2Client = new Oauth2Client();
+        OAuth2Appservice oauth2AppService = new OAuth2Appservice();
         String accessToken = "accessToken";
 
-        String email = oauth2Client.getEmail(accessToken);
+        String email = oauth2AppService.getEmail(accessToken);
 
         assertEquals(email, "email");
     }
